# quick-cart-backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quick-Cart Multi-Vendor E-Commerce API
  description: |
    # Quick-Cart API Documentation

    Welcome to the Quick-Cart Multi-Vendor E-Commerce Platform API. This API enables you to build powerful e-commerce applications with multi-vendor support, complete with role-based access control.

    ## 🔐 Authentication

    To test protected endpoints:
    1. Click the **"Authorize"** button (🔓 icon) at the top right
    2. Login using POST `/api/v1/auth/login` to get your JWT token
    3. Enter the token in the format: `Bearer <your_token_here>`
    4. Click **"Authorize"** to apply
    5. Now you can test all protected endpoints!

    ### Test Credentials
    Create an account first, or ask admin for credentials.

    ## 🎭 Roles & Permissions

    | Role | Description | Key Permissions |
    |------|-------------|-----------------|
    | **SUPER_ADMIN** | Platform administrator | All permissions |
    | **VENDOR** | Store owner/seller | Manage own products, view own orders |
    | **CUSTOMER** | Regular user | Browse, purchase, manage own orders |
    | **SUPPORT** | Customer support | View users, orders, products |

    ## 📝 Permission System

    Each endpoint shows required permissions in its description:
    - 🔓 **Public** - No authentication required
    - 🔒 **Protected** - Authentication required
    - 🛡️ **Permission Required** - Specific permission needed
    - 👤 **Ownership** - Must own the resource

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: `Bearer <token>`

        **How to get a token:**
        1. Register or login using the auth endpoints
        2. Copy the `accessToken` from the response
        3. Click "Authorize" button above
        4. Paste token and click "Authorize"
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [SUPER_ADMIN, VENDOR, CUSTOMER, SUPPORT]
          example: "CUSTOMER"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        phoneNumber:
          type: string
          nullable: true
          example: "+1234567890"
        isApproved:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/user/get-users:
    get:
      tags:
        - 👥 Users
      summary: Get all users
      description: |
        🛡️ **Required Permission:** `USER_READ`

        Retrieve a list of all users in the system.

        **Accessible by:**
        - SUPER_ADMIN ✅
        - SUPPORT ✅
        - Others ❌
      operationId: getAllUsers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: ✅ Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          description: ❌ Not authenticated
        "403":
          description: ❌ Insufficient permissions
  /api/v1/user/{userId}:
    get:
      tags:
        - 👥 Users
      summary: Get user by ID
      description: |
        🔒 **Protected** - Authentication required

        Get detailed information about a specific user.
        Users can view their own profile, admins can view anyone.
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
          example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: ✅ User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/User"
        "404":
          description: ❌ User not found
  /api/v1/user/{email}:
    put:
      tags:
        - 👥 Users
      summary: Update user profile
      description: |
        👤 **Ownership Required** - Can only update own profile

        Update user profile information. Users can only update their own data.
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
          example: "john@example.com"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phoneNumber:
                  type: string
                address:
                  type: string
                avatar:
                  type: string
      responses:
        "200":
          description: ✅ User updated successfully
  /api/v1/user/forgot-password:
    post:
      tags:
        - 👥 Users
      summary: Request password reset
      description: |
        🔓 **Public Endpoint**

        Request a password reset code. A security code will be sent to the email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john@example.com"
      responses:
        "200":
          description: ✅ Security code sent
        "404":
          description: ❌ User not found

  # ==================== PRODUCTS ====================
...